[wip] feat: models for problem and category.
